+++ JuliaBUGS/test/BUGSPrimitives/distributions.jl
@@ -15,9 +15,10 @@ end
         A[1:2, 1:2] ~ dwish(B[:, :], 2)
         C[1:2] ~ dmnorm(mu[:], A[:, :])
     end
-    model = compile(model_def, (mu=[0, 0], B=[1 0; 0 1]), (A=[1 0; 0 1],))
+    ad_model = compile(
+        model_def, (mu=[0, 0], B=[1 0; 0 1]), (A=[1 0; 0 1],); adtype=AutoReverseDiff()
+    )
 
-    ad_model = ADgradient(:ReverseDiff, model)
     theta = [
         0.7931743744870574,
         0.5151017206811268,